+++ cpp/turbodbc/Library/src/cursor.cpp
@@ -14,10 +14,14 @@
 
 namespace turbodbc {
 
-cursor::cursor(std::shared_ptr<cpp_odbc::connection const> connection, std::size_t rows_to_buffer, std::size_t parameter_sets_to_buffer) :
+cursor::cursor(std::shared_ptr<cpp_odbc::connection const> connection,
+               std::size_t rows_to_buffer,
+               std::size_t parameter_sets_to_buffer,
+               bool use_async_io) :
 	connection_(connection),
 	rows_to_buffer_(rows_to_buffer),
 	parameter_sets_to_buffer_(parameter_sets_to_buffer),
+	use_async_io_(use_async_io),
 	query_()
 {
 }
@@ -30,7 +34,7 @@ void cursor::prepare(std::string const & sql)
 	query_.reset();
 	auto statement = connection_->make_statement();
 	statement->prepare(sql);
-	query_ = std::make_shared<query>(statement, rows_to_buffer_, parameter_sets_to_buffer_);
+	query_ = std::make_shared<query>(statement, rows_to_buffer_, parameter_sets_to_buffer_, use_async_io_);
 }
 
 void cursor::execute()