+++ cpp/turbodbc/Library/turbodbc/query.h
@@ -15,7 +15,8 @@ class query {
 public:
 	query(std::shared_ptr<cpp_odbc::statement const> statement,
 		  std::size_t rows_to_buffer,
-		  std::size_t parameter_sets_to_buffer);
+		  std::size_t parameter_sets_to_buffer,
+		  bool use_double_buffering);
 
 	void execute();
 	std::shared_ptr<turbodbc::result_sets::result_set> get_results();
@@ -37,6 +38,7 @@ private:
 	std::shared_ptr<cpp_odbc::statement const> statement_;
 	std::size_t rows_to_buffer_;
 	std::size_t parameter_sets_to_buffer_;
+	bool use_double_buffering_;
 	std::vector<std::shared_ptr<turbodbc::parameter>> parameters_;
 	std::shared_ptr<result_sets::result_set> results_;
 	std::size_t current_parameter_set_;