+++ cpp/turbodbc/Library/turbodbc/result_sets/bound_result_set.h
@@ -20,7 +20,18 @@ public:
 	 *        the given statement.
 	 */
 	bound_result_set(std::shared_ptr<cpp_odbc::statement const> statement, std::size_t buffered_rows);
+
+	/**
+	 * @brief Move constructor
+	 */
+	bound_result_set(bound_result_set && other);
+
 	virtual ~bound_result_set();
+
+	/**
+	 * @brief Rebind the internal buffers to the statement.
+	 */
+	void rebind();
 private:
 	std::size_t do_fetch_next_batch() final;
 	std::vector<column_info> do_get_column_info() const final;