+++ python/turbodbc/connect.py
@@ -12,7 +12,7 @@ def _make_connection_string(dsn, **kwargs):
 
 
 @translate_exceptions
-def connect(dsn=None, rows_to_buffer=None, parameter_sets_to_buffer=None, **kwargs):
+def connect(dsn=None, rows_to_buffer=None, parameter_sets_to_buffer=None, use_async_io=False, **kwargs):
     """
     Create a connection with the database identified by the dsn
     :param dsn: data source name as given in the odbc.ini file
@@ -23,6 +23,11 @@ def connect(dsn=None, rows_to_buffer=None, parameter_sets_to_buffer=None, **kwar
                            shall be buffered when bulk queries are sent
                            to the server.
                            None means that the default value is used.
+    :param use_async_io: Set this value to true if you want to use
+                         asynchronous I/O. Asynchronous I/O means that
+                         while the main thread converts a batch of
+                         results to Python objects, another thread
+                         fetches additional results in the background.
     :param \**kwargs: You may specify additional options as you please.
                       These options will go into the connection string.
                       Valid options depend on the specific database you
@@ -35,5 +40,7 @@ def connect(dsn=None, rows_to_buffer=None, parameter_sets_to_buffer=None, **kwar
 
     if parameter_sets_to_buffer:
         connection.impl.parameter_sets_to_buffer = parameter_sets_to_buffer
+
+    connection.impl.use_async_io = use_async_io
 
     return connection